+++ js/util.js
@@ -1,0 +1,11 @@
+const getRandomInteger = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+const idGenerator = () => {
+  let lastGeneratedId = 0;
+  lastGeneratedId += 1;
+  return lastGeneratedId;
+};
+
+export {getRandomInteger, getRandomArrayElement, idGenerator};
